+++ app/session/store/memory.js
@@ -1,5 +1,11 @@
+/**
+ * In-memory HTTP session store.
+ *
+ */
 exports = module.exports = function() {
   var MemoryStore = require('express-session').MemoryStore;
   
   return new MemoryStore();
 };
+
+exports['@singleton'] = true;